Fixes to SparkManage, SparkQPTest

+++ src/main/java/cubemanager/sparkschema/SparkManager.java
@@ -221,7 +221,8 @@ public class SparkManager {
 				if(StringUtils.isNumeric(queryList.get(i).get(j).toString().replace(",", "")) || StringUtils.isNumeric(queryList.get(i).get(j).toString().replace(".", ""))) {
 					if (!StringUtils.isNumeric(queryList.get(i).get(j).toString())) {
 						DecimalFormat decim = new DecimalFormat("0.0000");
-						value = decim.format(Float.parseFloat(queryList.get(i).get(j).toString()))+"";
+						//value = decim.format(Float.parseFloat(queryList.get(i).get(j).toString()))+"";
+						value = decim.format(Double.parseDouble(queryList.get(i).get(j).toString()))+"";
 						value = value.replace(",", ".");
 					} else {
 						value = queryList.get(i).get(j).toString();
